Refactor and reorganize code
