+++ perf/gopool/worker.go
@@ -79,3 +79,59 @@ func newTypedTaskRunner[T any](handler func(context.Context, T)) taskRunner {
 		t.Recycle()
 	}
 }
+
+var taskPool sync.Pool
+
+type task struct {
+	ctx context.Context
+	arg any
+
+	next *task
+}
+
+func newTask() *task {
+	if t := taskPool.Get(); t != nil {
+		return t.(*task)
+	}
+	return &task{}
+}
+
+func (t *task) Recycle() {
+	*t = task{}
+	taskPool.Put(t)
+}
+
+type taskList struct {
+	mu    sync.Mutex
+	count int32
+	head  *task
+	tail  *task
+}
+
+func (l *taskList) add(t *task) (count int) {
+	l.mu.Lock()
+	if l.head == nil {
+		l.head = t
+		l.tail = t
+	} else {
+		l.tail.next = t
+		l.tail = t
+	}
+	l.count++
+	count = int(l.count)
+	l.mu.Unlock()
+	return
+}
+
+// pop acquired the lock and returns a task from the head of the taskList.
+//
+// Note that the caller takes responsibility to release the lock.
+func (l *taskList) pop() (t *task, lock *sync.Mutex) {
+	l.mu.Lock()
+	if l.head != nil {
+		t = l.head
+		l.head = l.head.next
+		l.count--
+	}
+	return t, &l.mu
+}
